zynqmp_pmufw: Fix violations for MISRA Rule 12_2
This patch fix following violation
misra_c_2012_rule_12_2_violation: The right hand operand of a shift
operator shall lie in the range zero to one less than the width in
bits of the essential type of the left hand operand.

To solve this MISRA violation do explicit type cast as per requirement.


Acked-for-series: Izhar Ameer Shaikh <izhar.ameer.shaikh@xilinx.com>

+++ lib/sw_apps/zynqmp_pmufw/src/pm_clock.c
@@ -622,7 +622,7 @@ static s32 PmClockGenSetDivider(PmClock* const clock, const u32 divId,
 		status = XST_INVALID_PARAM;
 		goto done;
 	}
-	XPfw_RMW32(clk->ctrlAddr, PM_DIV_MASK << shift, val << shift);
+	XPfw_RMW32(clk->ctrlAddr, (u32)PM_DIV_MASK << shift, (u32)val << shift);
 
 done:
 	return status;
@@ -779,8 +779,8 @@ static s32 PmClockCrossDomSetDivider(PmClock* const clock, const u32 divId,
 		status = XST_INVALID_PARAM;
 		goto done;
 	}
-	XPfw_RMW32(clk->ctrlAddr, PM_DIV_MASK << PM_DIV0_SHIFT,
-		   val << PM_DIV0_SHIFT);
+	XPfw_RMW32(clk->ctrlAddr, (u32)PM_DIV_MASK << PM_DIV0_SHIFT,
+		   (u32)val << PM_DIV0_SHIFT);
 
 done:
 	return status;
