+++ lib/sw_apps/zynqmp_pmufw/src/pm_pinctrl.c
@@ -1063,7 +1063,7 @@ s32 PmPinCtrlSetParam(const u32 pinId, const u32 paramId, const u32 value)
 	}
 
 	if (0U == (PM_PIN_PARAM_2_BITS & pmPinParams[paramId].flags)) {
-		XPfw_RMW32(addr, 1U << shift, value << shift);
+		XPfw_RMW32(addr, (u32)1 << shift, (u32)value << shift);
 		/* When setting pull up/down we need to enable pull as well */
 		if (paramId == PINCTRL_CONFIG_PULL_CTRL) {
 			addr = PM_PIN_PARAM_GET_ADDR(pinId,
@@ -1072,12 +1072,12 @@ s32 PmPinCtrlSetParam(const u32 pinId, const u32 paramId, const u32 value)
 			if (addr == IOU_SLCR_BANK1_CTRL5) {
 				FIX_BANK1_CTRL5(shift);
 			}
-			XPfw_RMW32(addr, 1U << shift, value << shift);
+			XPfw_RMW32(addr, (u32)1 << shift, (u32)value << shift);
 		}
 	} else {
 		/* Write value[0] at address + 4 and value[1] at address */
-		XPfw_RMW32(addr + 4U, 1U << shift, (value & 0x1U) << shift);
-		XPfw_RMW32(addr, 1U << shift, ((value & 0x2U) >> 1U) << shift);
+		XPfw_RMW32(addr + 4U, (u32)1 << shift, (value & 0x1U) << shift);
+		XPfw_RMW32(addr, (u32)1 << shift, ((value & 0x2U) >> 1U) << shift);
 	}
 
 done:
